Refresh MacOS support
I have a Big Sur (developer beta) shell. So just
updated a note in the shell setup. ZSH is now
the default shell in MacOS, so I purged a few
integrations related to OhMyZSH.

I want a more vanilla ZSH experience to boot if
I no longer use Bash. Bash has also seen updates,
and I still assume the user is using Bash 5 over the
vanilla Bash 3.2 shell.

The ChunkWM config remains as of yet useless,
as it does not run on the new MacOS.