🐟 Switch to Fish shell
It's about the time for a better interactive
shell experience. We can do better than Apple,
we can do better than ZSH. The "leap year" of
the shell is best modeled by Fish.

I still use Bash, but want to use it only
secondarily. I used Bash from 2016-2020
primarily, only because of wanting to keep
memory fresh of how to work with legacy
systems with only a POSIX/Bash console.

 * Mod+Enter now spawns fish shells.
 * Mod+Shift+Enter added for default
   (99.99% Bash in today's systems)

Fish is better most immediately because
of how it deals with aliases. The `abbr`
feature alone makes Fish compelling for
the daily shell user.

Added a `_setup` script which for now,
only contains the fisher setup.