Port `dir` to Fish
DIR

The configuration format for dir is
changed to be shareable between the
two shells, Bash and Fish.

Enough code is scaffolded that it
works in both shells, but more of
the innards should be ported to
Fish over time.

POWER ALIASES

There is a convention of putting these in
/sbin, and not adding /sbin to $PATH.

These aliases will make the /sbin utils
more reachable.