Switch to Fish on the Mac
I had to hack this by invoking Fish
in the bash profile for Mac since
`chsh -s $(which fish)` is blocked
in the Apple platform.

It is probably a safe default for
most users, but presents issues for
the developer.

Thus, there is a dangling Bash shell
for every instance of Fish (tradeoff).

I await a better solution.