Super
Remapped Caps lock to another Windows key.

GIVING LEFT PINKY WORK TO DO

Resetting keyboard layout without restarting:

 $ setxkbmap -layout us \
    -variant intl

Ref: https://askubuntu.com/a/1026171/456949

---------------------------------------------

In a [ LEFT RIGHT ] arrangement where LEFT is
the primary device...

Using a Group of 2 and 2,

 [ 1 2 ] [ 3 4 ]

with buckets [ 5 6 ] on reserve for additional
space, where 5 appears on the PRIMARY device,
the LEFT monitor in this arrangement.